Node上で動くことと、socket.ioの動作を確認
