+++ client/Map/map2.json
@@ -1,0 +1,1 @@
+[[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,122,122,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,158,158,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,122,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,100,102,0,0,100,102,0,0,0,0,0,0,0,0,0,0,158,0,1,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,0,0,124,126,0,0,124,126,0,0,0,0,0,0,0,122,122,0,0,0,0,0,0,0,0,0,0,0,0,0],[0,0,0,0,0,0,0,0,0,0,100,101,137,138,0,0,136,138,101,102,0,0,0,0,0,158,158,0,0,0,0,0,0,1,100,101,101,101,101,102,150],[100,101,101,101,101,101,101,101,101,101,124,125,137,138,0,0,136,138,125,126,101,101,101,102,0,0,0,0,0,0,0,0,0,0,124,125,125,125,125,126],[124,125,125,125,125,125,125,125,125,125,136,137,137,138,0,0,136,138,137,150,125,125,125,126,0,0,0,0,0,0,0,0,0,0,136,149,149,149,150,138],[136,137,137,137,137,137,137,137,137,137,136,137,137,138,0,0,136,138,137,150,149,149,149,150,0,0,0,0,0,0,0,0,0,1,136,149,149,149,150,138]]
