--- mapmake/index.html
@@ -1,48 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="ja">
-	<head>
-		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-		<title>Mapmake</title>
-		<link rel="stylesheet" type="text/css" href="./layout.css">
-		<script type = "text/javascript" src="mapmake.js"></script>
-	</head>
-	<body>
-		<div class="body_right">
-			<H2>MapMake</H2>
-			<table border=0 >
-			<tr><td>
-			<canvas id="img_list" width="384" height="768">
-				canvas機能を使用するためにはHTML5に対応したブラウザが必要になります。
-			</canvas>
-			</td><td>
-			<canvas id="now_img" width="32" height="32">
-				canvas機能を使用するためにはHTML5に対応したブラウザが必要になります。
-			</canvas>
-			</td><td>
-			<div id="now_select">
-			</div>
-			<canvas id="game_main" width="640" height="480">
-				canvas機能を使用するためにはHTML5に対応したブラウザが必要になります。
-			</canvas>
-			</td></tr>
-			</table>
-			<table border=1 class="ts">
-				<tr>
-					<td colspan=2>操作説明</td>
-				</tr>
-				<tr>
-					<td>左右キー</td>
-					<td>自機の左右移動</td>
-				</tr>
-				<tr>
-					<td>Aキー</td>
-					<td>ジャンプ</td>
-				</tr>
-				<tr>
-					<td>それ以外</td>
-					<td>自機を止める</td>
-				</tr>
-			</table>
-		</div>
-	</body>
-</html>
